requirements.txt
ccxt==4.3.92
google-cloud-bigquery==3.21.0
python-logging-loki==0.3.1
PyYAML==6.0.2
pydantic==2.8.2
schedule==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1

